Update symbols file for sparc64.
Forgot to check sparc64, powerpcspe and m68k (m68k is not built).
alpha and sparc64 (1.12.1-3.2) did not have the symbols *_32 or *_64.
All the *_32 symbols got bumped to 1.13.7 but some of the *_64 are 1.6.2.
When sparc64 gets built in experimental if the *_64 symbols appear I guess
they get bumped to 1.13.7 and this reverted.
.
Git-Dch: Ignore